clients: Fixed ivi_surface_listener event handler
It is illegal to have a NULL pointer mentioned for an interface
event handler. Having a NULL in place of a function pointer would
seg fault the client if the compositor were to send this event.

Fixed ivi_surface_listener to have a valid function (even though
it is empty). The only reason why clients like simple-egl were not
seg faulting today is because the ivi shell today is not calling
ivi_surface_send_visibility from anywhere today.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>