ivi-shell: bugfix, display new invoked application on UI correctly.
Creation of ivi_surface is not proper trigger to show new application
because ivi_surface can be created before the application draws the
content. Move this code to notification of configure. If it is new
application which have not been added to layers yet, add it to
application layer and call swith_mode to show it on UI.

Signed-off-by: Nobuhiko Tanibata <ntanibata@jp.adit-jv.com>