ivi-shell: Support clipping of ivi-surface and ivi-layer
Transformation of surface is implematened in update_prop() by several
sub-method. To calculate clipping region; clip_x/y and
clip_width/height, new submethod, calc_clipping_rectangle(), is added.
In this method, clipping region is calculated based on both source rectangle
of ivi_surface and the one of ivi_layer the ivi_surface belongs to. This
is because if source rectangle of ivi_surface is larger than the one of
ivi_layer, the clip region should be fit to ivi_layer.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>