ivi-layout: ivi-layout: implement surface clipping
View clip region is set in surface-local coordinates. To compute that
region, both the ivi-surface destination rectangle on an ivi-layer and
the ivi-layer destination rectangle in the global coordinates are
transformed back into the surface-local coordinates.

The transformations are computed by first forming the transformation
matrices for the forward mappings, and then inverting them. The inverse
matrices are used to transform the destination rectangles to the
surface-local coordinate system. The intersection of the rectangles is
the view clip mask.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>
Acked-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>