[updated]ivi-shell:  clipping feature review No.5
> > Document in what coordinate space the boundingbox must be given.

> > A more step-wise description might be better, e.g.
> > "Transform rect_input by the inverse of matrix, intersect with boundingbox, and store the result in rect_output."

commnet: Aug-17-2015
Talking about surface-local coordinate space when the function does not
use a weston_surface or weston_view or in any other way relate to the
surface-local coordinate space is a little confusing. It might be better
to say the boundingbox is in the same coordinate space as rect_output.
That way you don't have to explain how the matrix relates to the
surface-local coordinate space (and this function does not really care
about it, either, only callers do).

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>