ivi-shell: multi screens support
output->(x,y) has a coordinates of a screen a layer now belongs to. Layer
coordinate shall transform itself to global coordinates. This is
calucuated as matrix:m as well.

Additionally, according to ivi-shell specification, region of surface
outside of screen shall not be displayed. So, screen_dest_rect and
layer_dest_rect_in_screen are intersected to be clipped outside region.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>