[patch 07/14 --part1] Revert "ivi-hmi-controller: support protocol to get the number of screens"
This reverts commit 56f1d4243be49c81835a0bf156b11317fadb4d5e.

clients/ivishell-user-interface.cで作成するscreen数の取得方法を変更する。

wl_outputイベントをscreen数分受信できるので、get_screenリクエストおよび、
screen_numイベントで取得する方法を破棄する。

Comment:

is the request get_screen_num really needed?

The compositor could just send the screen_num event automatically on
binding the ivi_hmi_controller global. That would be less IPC traffic
and simpler protocol. Plus, main() in ivi-shell-user-interface.c is
already set up with the correct roundtrips to ensure the event is
handled before startup continues, if the event is triggered by the
binding.

Is it enough to get the number of ivi-screens? If you ever intend to
support output hotplug hmi-controller, getting just the number is not
enough, since every hotplug or unplug would rewrite the ivi-screen
indices. This is also a problem with the ivi-layer id scheme you use to
assign ivi-layers to ivi-screens. OTOH, this is only the demo in Weston
upstream, so it is ok for now.

I see the protocol is extended without proper versioning. For public or
stable protocols this is forbidden, but since this protocol extension is
private and its only users are in the weston tree and no external users
should exist, I can let this pass.

In summary: I would propose to remove the get_screen_num request. That
is all.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>