[patch 14/14 --fixes]
Comment 1:

> 1a0b54c ivi-shell: hmi-controller supports multi screens at
> random-mode.
>
> 	It is randomizing only inside the first screen, but sure, it is
> 	ok.
> 	Reviewed-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>

Additionally, surface_x/y is calcurated without surface_width/height.
This means a ivi_surface will be potentially overlapped to outside of
the screen. This is used to check that outside of a screen shall not be
displayed. To verify this behavior, ivi_layer shall be created with the
size more then actual size of screen as well.

Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>