[to be fixed] ivi-shell: hmi-controller creates layers whose size is larger than one of screen.
This allows it to locate a ivi_surface or part of ivi-surface outside of screen.
In ivi usecase, it shall not be displayed if the surface is overlapped
to the next screen. This code can test the usecase with mode: random
because surface_x,y can be set to whole a layer without
surface_width/height compared to before.

Signed-off-by: Nobuhiko Tanibata <nobuhiko_tanibata@xddp.denso.co.jp>
Reviewed-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>